.history/project/requirements_20251109163558.txt
aws-cdk-lib==2.215.0
constructs>=10.0.0,<11.0.0
aws-cdk.aws-apigatewayv2-alpha==2.215.0a0
aws-cdk.aws-apigatewayv2-integrations-alpha==2.215.0a0
